tester corrections

+++ TP3/Ej1/Experimentos/testFibonacci.cpp
@@ -103,10 +103,11 @@ int main() {
             nuevas.push_back({qj, uj, vj});
         }
 
-        int tiempo = medirTiempo(s, t);
+        double tiempo = medirTiempo(s, t);
 
         output_file << N << "," << tiempo << "\n";
     }
 
     return 0;
 }
+
